Fixing incorrect job information for queued jobs
If a job has the status "queued" during the "get_job" call, it is added to "self.jobs_tracking" with the status "queued". If you now call "get_job" again, no new job details are retrieved via the REST API, but only the old job details are returned. This is not a problem for completed and failed jobs, but it is for queued jobs. Therefore, it must be additionally checked that the job status is not "queued".