fix: Rename SQL Query to Data Engine where appropriate
The IBM Cloud SQL Query service is now "Data Engine". This needs
renaming of e.g. API Endpoints and docs and code. This commit does
the renaming on all places where reasonable without breaking anything.

Note that "SQL Query" is used in 3 flavors in this repo:
1) As a term for the IBM Cloud service
2) As the actual query to get data from the service
3) As identifier, e.g. in URLs and code

This commit only changes the first flavor, and the needed parts
of the third one, e.g. for the API endpoints. The rest is kept
as is.